pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-digest"
version = "0.1.0"
description = "Daily AI digest for Telegram + Obsidian"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
  "fastapi>=0.116",
  "uvicorn>=0.30",
]

[project.optional-dependencies]
runtime = [
  "feedparser>=6.0.11",
  "beautifulsoup4>=4.12",
]
test = [
  "pytest>=8.0",
]

[project.scripts]
digest = "digest.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
